Import keys handle combine, override, cancel

+++ extension/src/Pages/ImportExportKeys.js
@@ -1,10 +1,13 @@
-import React from "react";
+import React, { useState } from "react";
 import Button from "react-bootstrap/Button";
 import { BsDownload } from "react-icons/bs";
 import Form from "react-bootstrap/Form";
 import Alert from "react-bootstrap/Alert";
+import { confirmAlert } from "react-confirm-alert"; // Import
+import "react-confirm-alert/src/react-confirm-alert.css"; // Import css
 
 function ImportExportKeys() {
+	const [importSuccess, setImportSuccess] = useState(false);
 	const exportKeys = () => {
 		const keys = JSON.parse(localStorage.getItem("Keys"));
 		const element = document.createElement("a");
@@ -20,7 +23,50 @@ function ImportExportKeys() {
 		const reader = new FileReader();
 		reader.onload = () => {
 			const importedKeys = JSON.parse(reader.result);
-			localStorage.setItem("Keys", JSON.stringify(importedKeys));
+			if (localStorage.getItem("Keys")) {
+				confirmAlert({
+					title: "Existing Keys Found",
+					message: `How would you like to handle this??`,
+					buttons: [
+						{
+							label: "Override",
+							onClick: () => {
+								localStorage.setItem("Keys", JSON.stringify(importedKeys));
+								setImportSuccess(true);
+								setTimeout(() => {
+									setImportSuccess(false);
+								}, 3000);
+							},
+						},
+						{
+							label: "Combine",
+							onClick: () => {
+								//TODO: handle duplicates also
+								const existingKeys = JSON.parse(localStorage.getItem("Keys"));
+								for (let website in importedKeys) {
+									for (let key_name in importedKeys[website]) {
+										existingKeys[website][key_name] = importedKeys[website][key_name];
+									}
+								}
+								console.log("updated keys comb", existingKeys);
+								// localStorage.setItem("Keys", JSON.stringify(existingKeys));
+								setImportSuccess(true);
+								setTimeout(() => {
+									setImportSuccess(false);
+								}, 3000);
+							},
+						},
+						{
+							label: "Cancel",
+							onClick: () => {
+								return;
+							},
+						},
+					],
+				});
+			}
+			// reset the input field
+			event.target.value = "";
 		};
 		reader.readAsText(file);
 	};
@@ -36,10 +82,10 @@ function ImportExportKeys() {
 
 			<div className='my-5 pb-3 pt-2' style={{ display: "flex", flexDirection: "column", alignItems: "center", justifyContent: "center" }}>
 				<h4>Import Keys</h4>
-				{/* <input type='file' accept='application/json' onChange={importKeys} /> */}
 				<Form.Group controlId='formFile' className='mb-3'>
 					<Form.Control type='file' accept='application/json' onChange={importKeys} />
 				</Form.Group>
+				{importSuccess && <Alert variant='success'>Keys Imported Successfully</Alert>}
 			</div>
 
 			<Alert variant='warning' className='mx-3'>
